Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Do you really need review on a new initial empty commit? :D

Not really, no
 
> Note that in general though I'm against the idea unless you can justify existing yaml libs don't suit you, did you properly try all of libyaml, libcyaml, libfyaml ? (last two aren't packaged on fedora :/ so no epel for free..)

If you take a look at the README in the last patch set, you will see that MiniYAML is not a completely new YAML library. It is just a few wrappers over libyaml. So everything is fine =)

> Parsing things in C is hard, so maybe it's possible to use libyaml-cpp or the rust implementations from c instead? Maintaining bindings might be less work in the long run.

I did not have a look at libyaml-cpp nor the rust implementation. I must admit I was a bit disappointed with the speed at which libyaml parses things, so maybe later I can give those a try.

Patch-set: 1