Update patch set 1
Patch Set 1:

Dominique is right, reviewing this patch (itself) is pretty meaningless.

Maybe we can use it as a platform to address any concern with:

- the name of the repository ;
- the point of maintaining miniyaml as a separate project (as opposed to just a part of librobinhood) ;
- the point of discussing any of the previous points now that I created the repo on github ;-) ;
- anything else that comes to your mind about the project as a whole.

To answer the second point (which Dominique mentioned before). I would like to maintain miniyaml as a separate project/repository because I feel like libyaml is not dev-friendly enough as it is.

I think this might appeal to a larger community than just robinhood's. For example, we could post it as a "Show HN" on hacker news. (We can discuss this elsewhere if you want.)

Patch-set: 1